refactor: ResultLottoStatistics package 위치 변경
