+++ src/main/kotlin/lotto/vo/ResultLottoStatistics.kt
@@ -1,4 +1,7 @@
-package lotto.domain
+package lotto.vo
+
+import lotto.domain.LottoProfitRate
+import lotto.domain.WinningLottoStatistics
 
 class ResultLottoStatistics(
     val winningLottoStatistics: WinningLottoStatistics,
